linux i386: find root partition for UKIs
No EFI variables in BIOS mode, so no way to inform
systemd-gpt-auto-generator about the location of the ESP,
to allow it finding the root partition itself.
If no root device is specified in the cmdline of the UKI:
scan the gpt for an x86_64 root partition (by part-type-uuid)
and add it to the cmdline via "root=PARTUUID=..."
In BIOS mode there is no secure boot so it is fine to modify the
cdmline.