bli: Add explicit dependency on the part_gpt module
The bli module has a "hidden" dependency on the part_gpt module,
which is not picked up automatically by the build system.
Without the part_gpt module, bli does not function correctly.
Since bli does its work in the module initialization function,
the order in which the modules are loaded is also important:
part_gpt needs to be loaded before the bli module.

To solve this, track this dependency explicitly.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>